comment about code for back and forth
In a quiz code is required to ensure that previously selected values are maintained if the user moves
back and forth. So if you select checkboxes one and three and then scroll to the next question without 
submitting. If you then scroll back to the question the user will expect those selections to still be present.